BACKEND/ML/requirements.txt
# Data Processing and Analysis
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
imbalanced-learn==0.11.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# ML Ops
mlflow==2.7.0
joblib==1.3.1
pyyaml==6.0
dagshub

# API
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6

# Cloud and Storage
boto3==1.28.15
s3fs
# Utilities
python-dotenv==1.0.0
s3fs 
pydantic<2.0.0
evidently==0.4.25
langchain-core==0.1.7
ensure==1.0.3
python-box==7.1.1
chardet==5.2.0 